+++ flag/service/prometheus/prometheus.go
@@ -4,9 +4,15 @@ type Prometheus struct {
 	Address    string
 	BaseDomain string
 	Storage    PrometheusStorage
+	Retention  PrometheusRetention
 }
 
 type PrometheusStorage struct {
 	CreatePVC string
 	Size      string
+}
+
+type PrometheusRetention struct {
+	Duration string
+	Size     string
 }