+++ service/controller/resource/resource.go
@@ -17,6 +17,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/promxy"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/scrapeconfigs"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/servicemonitor"
+	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/volumeresizehack"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
@@ -29,6 +30,8 @@ type Config struct {
 	StorageSize             string
 	RestrictedAccessEnabled bool
 	WhitelistedSubnets      string
+	RetentionDuration       string
+	RetentionSize           string
 	K8sClient               k8sclient.Interface
 	Logger                  micrologger.Logger
 	PrometheusClient        promclient.Interface
@@ -87,14 +90,30 @@ func New(config Config) ([]resource.Interface, error) {
 	var prometheusResource resource.Interface
 	{
 		c := prometheus.Config{
-			Address:          config.Address,
-			PrometheusClient: config.PrometheusClient,
-			Logger:           config.Logger,
-			CreatePVC:        config.CreatePVC,
-			StorageSize:      config.StorageSize,
+			Address:           config.Address,
+			PrometheusClient:  config.PrometheusClient,
+			Logger:            config.Logger,
+			CreatePVC:         config.CreatePVC,
+			StorageSize:       config.StorageSize,
+			RetentionDuration: config.RetentionDuration,
+			RetentionSize:     config.RetentionSize,
 		}
 
 		prometheusResource, err = prometheus.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var volumeResizeHack resource.Interface
+	{
+		c := volumeresizehack.Config{
+			Logger:           config.Logger,
+			K8sClient:        config.K8sClient,
+			PrometheusClient: config.PrometheusClient,
+		}
+
+		volumeResizeHack, err = volumeresizehack.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -178,6 +197,7 @@ func New(config Config) ([]resource.Interface, error) {
 		apiCertificatesResource,
 		tlsCertificatesResource,
 		prometheusResource,
+		volumeResizeHack,
 		serviceMonitorResource,
 		alertResource,
 		scrapeConfigResource,